requirements.txt
livekit>=1.0
livekit-agents[openai]~=1.0
python-dotenv~=1.0
msgraph-sdk==1.5.4
azure-identity==1.19.0
requests>=2.31.0
msgraph-core==1.1.5
azure-core==1.29.5